api/requirement.txt

fastapi
uvicorn

pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy